[FLINK-35230][sql-parser] Split FlinkSqlParserImplTest into smaller test classes
