+++ flink-table/flink-sql-parser/pom.xml
@@ -36,7 +36,7 @@ under the License.
 		<!-- override parent pom -->
 		<test.excludedGroups/>
 		<surefire.module.config><!--
-			FlinkSqlParserImplTest has to override Calcite's SqlParserTest
+			FlinkSqlParserTestBase subclasses override SqlParserFixture methods
 			JUnit changed this behavior since 5.11.x, to restore it back need this flag as mentioned at
 			https://junit.org/junit5/docs/current/user-guide/#extensions-supported-utilities-search-semantics
 			-->-Djunit.platform.reflection.search.useLegacySemantics=true
