+++ tools/maven/suppressions.xml
@@ -34,7 +34,6 @@ under the License.
 		<suppress files="NoticeFileCheckerTest.java" checks="IllegalImport"/>
 		<suppress files="DependencyTree.java" checks="IllegalImport"/>
 
-		<suppress files="FlinkSqlParserImplTest.java" checks="FileLength"/>
 		<suppress files="JoinOperator.java" checks="FileLength"/>
 		<suppress files="WindowOperatorTest.java" checks="FileLength"/>
 		<suppress files="WindowOperatorContractTest.java" checks="FileLength"/>
